preprocess handle single end reads; clean out files;

+++ bin/TranslocPreprocess.pl
@@ -52,11 +52,14 @@ my $read2;
 my $indir;
 my $outdir;
 my $meta_file;
+my $paired_end = 1;
 my $max_threads = 4;
 my $bc_len = 10;
 my $bc_mismatch = 2;
+my $bc_min_dist = 2;
 my $adapt_max_dif = 20;
 my $minlen = 30;
+my $no_trim;
 my $join;
 my $join_max_dif = 10;
 my $join_min_ol = 30;
@@ -89,7 +92,10 @@ read_in_meta_file;
 unless (defined $indir) {
 	create_barcode_file;
 	mkdir "$outdir/multx";
-	my $multx_cmd = "fastq-multx -m $bc_mismatch -x -b -B $bcfile $read1 $read2 -o $outdir/multx/%_R1.fq.gz $outdir/multx/%_R2.fq.gz";
+	my $multx_cmd = join(" ", "fastq-multx -m $bc_mismatch -d $bc_min_dist -x -b -B $bcfile $read1",
+														$paired_end ? "$read2 -o $outdir/multx/%_R1.fq.gz $outdir/multx/%_R2.fq.gz" :
+																					"-o $outdir/multx/%.fq.gz");
+
 	@bc_output = Capture($multx_cmd) or croak "Error: fastq-multx failed";
 	print "\n@bc_output\n";
 } else {
@@ -103,49 +109,51 @@ mkdir "$outdir/trim";
 mkdir "$outdir/join";
 mkdir "$outdir/logs";
 
-my @threads = ();
+if ($paired_end) {
+	my @threads = ();
 
-foreach my $expt (sort keys %meta) {
+	foreach my $expt (sort keys %meta) {
 
-	while (1) {
+		while (1) {
 
-	# joins any threads if possible
-		foreach my $thr (@threads) {
-			$thr->join() if $thr->is_joinable();
-		}
+		# joins any threads if possible
+			foreach my $thr (@threads) {
+				$thr->join() if $thr->is_joinable();
+			}
 
-		my @running = threads->list(threads::running);
-		
-		# if there are open threads, create a new one, push it onto list, and exit while loop
-		if (scalar @running < $max_threads) {
-			my $thr = threads->create( sub {
-						my $t0_expt = [gettimeofday];
-						print "\nStarting $expt\n";
-						process_experiment($expt);
-						my $t1_expt = tv_interval($t0_expt);
-						printf("\nFinished %s with %d reads in %.2f seconds.\n", $expt, $meta{$expt}->{multx},$t1_expt);
-					});
-			push(@threads,$thr);
-			sleep(1);
-			last;
+			my @running = threads->list(threads::running);
+			
+			# if there are open threads, create a new one, push it onto list, and exit while loop
+			if (scalar @running < $max_threads) {
+				my $thr = threads->create( sub {
+							my $t0_expt = [gettimeofday];
+							print "\nStarting $expt\n";
+							process_experiment($expt);
+							my $t1_expt = tv_interval($t0_expt);
+							printf("\nFinished %s with %d reads in %.2f seconds.\n", $expt, $meta{$expt}->{multx},$t1_expt);
+						});
+				push(@threads,$thr);
+				sleep(1);
+				last;
+			}
+			sleep(5);
+		} 
+	}
+
+	# waits for all threads to finish
+	while( scalar threads->list(threads::all) > 0) {
+		for my $thr (@threads) {
+			$thr->join() if $thr->is_joinable;
 		}
 		sleep(5);
-	} 
-}
-
-# waits for all threads to finish
-while( scalar threads->list(threads::all) > 0) {
-	for my $thr (@threads) {
-		$thr->join() if $thr->is_joinable;
 	}
-	sleep(5);
+
+	unless (defined $indir) {}
+
+
+	update_stats_from_trim;
+
 }
-
-unless (defined $indir) {}
-
-
-update_stats_from_trim;
-
 
 write_stats_file;
 
@@ -194,8 +202,12 @@ sub create_barcode_file {
 	my $bcfh = IO::File->new(">$bcfile");
 
 	foreach my $expt (sort keys %meta) {
-		$meta{$expt}->{R1} = "$outdir/multx/${expt}_R1.fq.gz";
-		$meta{$expt}->{R2} = "$outdir/multx/${expt}_R2.fq.gz";
+		if ($paired_end) {
+			$meta{$expt}->{R1} = "$outdir/multx/${expt}_R1.fq.gz";
+			$meta{$expt}->{R2} = "$outdir/multx/${expt}_R2.fq.gz";
+		} else {
+			$meta{$expt}->{R1} = "$outdir/multx/$expt.fq.gz";
+		}
 		my $barcode = uc(substr($meta{$expt}->{mid}.$meta{$expt}->{primer},0,$bc_len));
 		$bcfh->print(join("\t",$expt,$barcode)."\n");
 		print "$expt $barcode\n";
@@ -209,8 +221,8 @@ sub check_existance_of_files {
 	foreach my $expt (sort keys %meta) {
 		$meta{$expt}->{R1} = "$indir/${expt}_R1.fq.gz";
 		$meta{$expt}->{R2} = "$indir/${expt}_R2.fq.gz";
-		croak "Error: Could not locate $expt read 1 file" unless ($meta{$expt}->{R1});
-		croak "Error: Could not locate $expt read 2 file " unless ($meta{$expt}->{R2});
+		croak "Error: Could not locate $expt read 1 file" unless (-r $meta{$expt}->{R1});
+		croak "Error: Could not locate $expt read 2 file " unless (-r $meta{$expt}->{R2});
 
 	}
 }
@@ -265,28 +277,32 @@ sub process_experiment ($) {
 
 	my $logfile = "$outdir/logs/$expt.log";
 
+
 	System("echo \"Pre-processing $expt\" > $logfile",1);
 
+	unless (defined $no_trim) {
 
-	$meta{$expt}->{R1trim} = "$outdir/trim/${expt}_R1.fq.gz";
-	$meta{$expt}->{R2trim} = "$outdir/trim/${expt}_R2.fq.gz";
+		$meta{$expt}->{R1trim} = "$outdir/trim/${expt}_R1.fq.gz";
+		$meta{$expt}->{R2trim} = "$outdir/trim/${expt}_R2.fq.gz";
 
-	System("echo \"Running SeqPrep on $expt\" >> $logfile",1);
+		System("echo \"Running SeqPrep on $expt\" >> $logfile",1);
 
-	my $t0_trim = [gettimeofday];
-						
+		my $t0_trim = [gettimeofday];
+							
 
-	my $trim_cmd = join(" ","SeqPrep -f",$meta{$expt}->{R1},"-r",$meta{$expt}->{R2},
-													"-1",$meta{$expt}->{R1trim},"-2",$meta{$expt}->{R2trim},
-													"-L $minlen -A $forward_adapter -B $reverse_adapter >> $logfile 2>&1");
+		my $trim_cmd = join(" ","SeqPrep -f",$meta{$expt}->{R1},"-r",$meta{$expt}->{R2},
+														"-1",$meta{$expt}->{R1trim},"-2",$meta{$expt}->{R2trim},
+														"-L $minlen -A $forward_adapter -B $reverse_adapter >> $logfile 2>&1");
 
-	System("echo \"$trim_cmd\" >> $logfile",1);
+		System("echo \"$trim_cmd\" >> $logfile",1);
 
-	System($trim_cmd,1) or croak "Error: failed running SeqPrep on $expt";
+		System($trim_cmd,1) or croak "Error: failed running SeqPrep on $expt";
 
-	my $t1_trim = tv_interval($t0_trim);
+		my $t1_trim = tv_interval($t0_trim);
 
-	System("echo \"Finished fastq-trim on $expt in $t1_trim seconds\" >> $logfile",1);
+		System("echo \"Finished fastq-trim on $expt in $t1_trim seconds\" >> $logfile",1);
+
+	}
 
 	if ($join) {
 		$meta{$expt}->{R1join} = "$outdir/join/${expt}_R1.fq.gz";
@@ -321,7 +337,7 @@ sub write_stats_file {
 	foreach my $expt (sort keys %meta) {
 		my @row = ($expt,
 			$meta{$expt}->{multx},
-			$meta{$expt}->{trim});
+			defined $meta{$expt}->{trim} ? $meta{$expt}->{trim} : "");
 		$statsfh->print(join("\t",@row)."\n");
 	}
 	unless (defined $indir) {
@@ -336,12 +352,15 @@ sub clean_up {
 
 	# Do not delete demultiplexed files - need for data submission
 	# System("rm $outdir/multx/*");
-
-	unless ($join) {
-		System("mv $outdir/trim/* $outdir/" );
+	if ($paired_end) {
+		unless ($join) {
+			System("mv $outdir/trim/* $outdir/" );
+		} else {
+			System("mv $outdir/join/* $outdir/");
+			System("rm $outdir/trim/*")
+		}
 	} else {
-		System("mv $outdir/join/* $outdir/");
-		System("rm $outdir/trim/*")
+		System("mv $outdir/multx/* $outdir/")
 	}
 
 }
@@ -356,14 +375,16 @@ sub parse_command_line {
   			"read2=s" => \$read2,
   			"indir=s" => \$indir,
 				"threads=i" => \$max_threads,
-				"bclen=i" => \$bc_len,
-				"bcmismatch=i" => \$bc_mismatch,
-				"adapt_max_dif=i" => \$adapt_max_dif,
+				"bc-len=i" => \$bc_len,
+				"bc-mismatch=i" => \$bc_mismatch,
+				"bc-min-dist=i" => \$bc_min_dist,
+				"trim-max-dif=i" => \$adapt_max_dif,
+				"no-trim" => \$no_trim,
 				"join" => \$join,
-				"join_max_dif=i" => \$join_max_dif,
-				"join_min_ol=i" => \$join_min_ol,
-				"minlen=i" => \$minlen,
-				"skipclean" => \$skipclean,
+				"join-max-dif=i" => \$join_max_dif,
+				"join-min-ol=i" => \$join_min_ol,
+				"trim-min-length=i" => \$minlen,
+				"skip-clean" => \$skipclean,
 				"help" => \$help
 		  ) ;
   
@@ -378,7 +399,9 @@ sub parse_command_line {
 		croak "Error: cannot define both input directory and non-de-multiplexed reads" if (defined $read1 || defined $read2);		
 	} else {
 		croak "Error: cannot find read 1 $read1 does not exist" unless (-r $read1);
-		croak "Error: cannot find read 2 $read2 does not exist" unless (-r $read2);
+		unless (-r $read2) {
+			$paired_end = 0;
+		}
 	}
 	unless (-d $outdir) {
 		mkdir $outdir or croak "Error: output directory $outdir neither exists nor could be created";
@@ -408,14 +431,14 @@ $arg{"--read1","Illumina read 1 file for de-multiplexing"}
 $arg{"--read2","Illumina read 2 file for de-multiplexing"}
 $arg{"--indir","Input directory - if specified assumes reads are already de-multiplexed"}
 $arg{"--threads","Number of threads to run program on - experiments processed simultaneously",$max_threads}
-$arg{"--bclen","Number of bases to use from primer file for de-multiplexing",$bc_len}
-$arg{"--bcmismatch","Number of mismatches allowed in de-multiplexing",$bc_mismatch}
+$arg{"--bc-len","Number of bases to use from primer file for de-multiplexing",$bc_len}
+$arg{"--bc-mismatch","Number of mismatches allowed in de-multiplexing",$bc_mismatch}
 $arg{"--adapter","Fasta file of adapter sequences"}
-$arg{"--adapt_max_dif","Maximum percent difference for match with adapter",$adapt_max_dif}
+$arg{"--trim-max-dif","Maximum percent difference for match with adapter",$adapt_max_dif}
 $arg{"--join","Stitch reads together using fastq-join"}
-$arg{"--join_max_dif","Maximum percent difference between reads for quality-trimmed stitching",$join_max_dif}
-$arg{"--join_min_ol","Minimum basepair overlap between two reads for quality-trimmed stitching",$join_min_ol}
-$arg{"--minlength","Minimum length of quality trimmed reads",$minlen}
+$arg{"--join-max-dif","Maximum percent difference between reads for quality-trimmed stitching",$join_max_dif}
+$arg{"--join-min-ol","Minimum basepair overlap between two reads for quality-trimmed stitching",$join_min_ol}
+$arg{"--trim-min-length","Minimum length of quality trimmed reads",$minlen}
 $arg{"--help","This helpful help screen."}
